Pulled out Helmholtz decomp into separate class.
Started testing it and realized that while solving
the poisson problem gives perfect velocity at the
boundaries, it does not give zero divergence. It's
close to zero, and that closeness goes up as the,
grid is larger, but it's still pretty imprecise.

The reason is that when we project w1 to w2:

w2 = w1 - grad P,

And then do a divergence, we get:

div w2 = div w1 - div grad P,

Now our poisson equation was: div w1 = div grad P,
and it was solved perfectly with 13 digits of,
precision. BUT, div grad P is NOT the same as
laplacian of P when discretized. Because div grad
involves two central differences, which ends up
not being the same as the 5 point laplacian
stencil.

In 1D, laplacian of P is P[i+1] + P[i-1] - 2P[i],
but div grad P = div (P[i+1] - P[i-1])/2 =
= ((P[i+2] - P[i])/2 - (P[i] - P[i-2])/2)/2 =
(P[i+2] + P[i-2] - 2P[i])/4.

These expressions have the same order of magnitude
of error in the taylor series, but with different
constants.